Refactor `header.rs`
This commit is part of a series of commits to refactor this crate to
support Android (0,1,2,3,4 versions) and Samsung boot image headers.

Changes will be required in the Image modules to handle this change.

Currently, the approach is to use traits to allow for a shared spec of
methods in each Header struct.